First commit
A css animation files are given in  the folder. On mouse hover the text start to animate and on removal it may be helpful for designing purpose or for those who want to learn animation effect with the use of html and css only. 
Hope you will like this.
Happy coding.