admin commit
The animation files are given in  the folder. On mouse hover the text content will display a different text with the effect of glowing. It may be helpful for designing purpose or for those who want to learn animation effect with the use of html and css only. 
Hope you will like this.
Happy coding.