admin commit
Hello guys, in this folder html and CSS files are given which will gave an animation of circle in dots. As the page load the dots start to move in a circular path which will add a factor to any website look. It may be helpful for designing purpose or for those who want to learn animation effect. 
Hope you will like this.
Happy coding.