admin commit
Hello guys, in this folder html and CSS files are given which will gave an animation of 3D cubes which are basically the gallery images and also on hover on these cubes they will start to rotate as for displaying all images and on removal it gets back to it's original position. It may be helpful for designing purpose or for those who want to learn animation effect. 
Hope you will like this.
Happy coding.